+++ src/extension.ts
@@ -11,6 +11,7 @@ import { CreateFlightplanHeader } from './Commands/flightplan/create-header';
 import { CreateFlightplanHeaderFromAifp } from './Commands/flightplan/create-header-from-aifp';
 import { GenerateAirports } from './Commands/flightplan/generate-airports';
 import { MatchAcNumbers } from './Commands/flightplan/match-ac-numbers';
+import { OpenMasterAirportsFile } from './Commands/open-master-airports-file';
 import { FlightplanMetadata } from './Commands/flightplan/metadata';
 import { RebaseAircraftNumbers } from './Commands/flightplan/rebase-ac-numbers';
 import { RenameFiles } from './Commands/flightplan/rename-files';
@@ -72,6 +73,10 @@ export function activate(context: vscode.ExtensionContext) {
 
 		vscode.commands.registerCommand('fsAiTools.matchAcNumbers', () => {
 			MatchAcNumbers();
+		}),
+
+		vscode.commands.registerCommand('fsAiTools.openMasterAirportsFile', () => {
+			OpenMasterAirportsFile();
 		}),
 
 		vscode.commands.registerCommand('fsAiTools.rebaseAircraftNumbers', () => {